+++ Content.Server/Chat/Systems/ChatSystem.cs
@@ -61,6 +61,7 @@ public sealed partial class ChatSystem : SharedChatSystem
     [Dependency] private readonly IReplayRecordingManager _replay = default!;
     [Dependency] private readonly IConfigurationManager _configurationManager = default!;
     [Dependency] private readonly IChatManager _chatManager = default!;
+    [Dependency] private readonly HearingSystem _hearing = default!;
     [Dependency] private readonly IChatSanitizationManager _sanitizer = default!;
     [Dependency] private readonly IAdminManager _adminManager = default!;
     [Dependency] private readonly IPlayerManager _playerManager = default!;
@@ -522,7 +523,7 @@ public sealed partial class ChatSystem : SharedChatSystem
                 continue; // Won't get logged to chat, and ghosts are too far away to see the pop-up, so we just won't send it to them.
 
             // WWDP Deafening
-            if (!ProcessDeafness(session, ChatChannel.Whisper))
+            if (_hearing.IsBlockedByDeafness(session, ChatChannel.Whisper, language))
                 continue;
             // WWDP end
 
@@ -741,7 +742,7 @@ public sealed partial class ChatSystem : SharedChatSystem
             var entHideChat = entRange == MessageRangeCheckResult.HideChat;
 
             // WWDP Deafening
-            if (!ProcessDeafness(session, channel))
+            if (_hearing.IsBlockedByDeafness(session, channel, language))
                 return;
             // WWDP end
 
@@ -782,27 +783,6 @@ public sealed partial class ChatSystem : SharedChatSystem
         }
         // WD EDIT END
     }
-
-    // WWDP Deafening
-    // Returns false if the message can not be heard
-    // Sends a DeafChatMessage in player's chat
-    private bool ProcessDeafness(ICommonSession session, ChatChannel channel)
-    {
-        if (channel is not (ChatChannel.Local or ChatChannel.Whisper or ChatChannel.Radio or ChatChannel.Notifications))
-            return true;
-
-        if (TryComp<DeafComponent>(session.AttachedEntity, out var deafComp))
-        {
-            var canthearmessage = Loc.GetString(deafComp.DeafChatMessage);
-            var wrappedcanthearmessage = $"{canthearmessage}";
-
-            _chatManager.ChatMessageToOne(ChatChannel.Local, canthearmessage, wrappedcanthearmessage, EntityUid.Invalid, false, session.Channel);
-            return false;
-        }
-
-        return true;
-    }
-    // WWDP end
 
     /// <summary>
     ///     Returns true if the given player is 'allowed' to send the given message, false otherwise.
